Casted ints to floats in parse.c. It now builds with gcc 5.4.0 on Ubuntu 16.04

+++ src/app/interfaces/iGPS/lib_nmea/src/parse.c
@@ -905,7 +905,7 @@ int nmea_parse_PRTI01(const char *s, const int len, bool has_checksum, nmeaPRTI0
     return 0;
   }
 
-  if (!isnan(pack->bottom_vx) && !isnan(pack->bottom_vy) && !isnan(pack->bottom_vz)) {
+  if (!isnan((float)pack->bottom_vx) && !isnan((float)pack->bottom_vy) && !isnan((float)pack->bottom_vz)) {
     nmea_INFO_set_present(&pack->present, RTI);
   }
